+++ server/db/update_v0.9.10_v0.9.11.sql
@@ -1,11 +1,14 @@
+-- Increase validation code length
 ALTER TABLE `validation_codes` CHANGE `code` `code` VARCHAR(16) CHARACTER SET utf8 COLLATE utf8_general_ci NOT NULL;
 
+-- Add email modification page
 INSERT INTO `pages` (`id`, `keyword`, `url`, `protocol`, `id_actions`, `id_navigation_section`, `parent`, `is_headless`, `nav_position`, `footer_position`, `id_type`) VALUES (NULL, 'email', '/admin/email/[i:id]?', 'GET|POST|PATCH', '0000000002', NULL, '0000000009', '0', '11', NULL, '0000000001');
 SET @id_page_email = LAST_INSERT_ID();
 
 INSERT INTO `pages_fields_translation` (`id_pages`, `id_fields`, `id_languages`, `content`) VALUES (@id_page_email, '0000000008', '0000000001', 'Email CMS');
 INSERT INTO `acl_groups` (`id_groups`, `id_pages`, `acl_select`, `acl_insert`, `acl_update`, `acl_delete`) VALUES ('0000000001', @id_page_email, '1', '0', '1', '0'), ('0000000002', @id_page_email, '1', '0', '1', '0');
 
+-- Add email default content
 INSERT INTO `fieldType` (`id`, `name`, `position`) VALUES (NULL, 'email', '90');
 SET @id_field_email = LAST_INSERT_ID();
 
@@ -24,6 +27,7 @@ INSERT INTO `pages_fields_translation` (`id_pages`, `id_fields`, `id_languages`,
 (@id_page_email, @id_field_email_reminder, '0000000002', 'Guten Tag\r\n\r\nSie waren für längere Zeit nicht mehr aktiv auf der @project Plattform.\r\nEs würde uns freuen wenn Sie wieder vorbeischauen würden.\r\n\r\n@link\r\n\r\nMit freundlichen Grüssen\r\nihr @project Team'),
 (@id_page_email, @id_field_email_reminder, '0000000003', 'Hello\r\n\r\nYou did not visit the @project platform for some time now.\r\nWe would be pleased if you would visit us again.\r\n\r\n@link\r\n\r\nSincerely, your @project team');
 
+-- fix typo
 UPDATE `sections_fields_translation` SET `content` = 'Benutzername' WHERE `sections_fields_translation`.`id_sections` = 0000000026 AND `sections_fields_translation`.`id_fields` = 0000000036 AND `sections_fields_translation`.`id_languages` = 0000000002 AND `sections_fields_translation`.`id_genders` = 0000000001;
 
 -- update user table to set a default language